create task to update system resources

+++ edge-agent/src/Microsoft.Azure.Devices.Edge.Agent.Core/metrics/SystemResources.cs
@@ -8,23 +8,23 @@ namespace Microsoft.Azure.Devices.Edge.Agent.Core.Metrics
 
     public class SystemResources
     {
-        public SystemResources(double usedRam, double totalRam, Disk[] disks)
+        public SystemResources(long usedRam, long totalRam, Disk[] disks)
         {
             this.UsedRam = usedRam;
             this.TotalRam = totalRam;
             this.Disks = disks;
         }
 
-        public double UsedRam { get; }
+        public long UsedRam { get; }
 
-        public double TotalRam { get; }
+        public long TotalRam { get; }
 
         public Disk[] Disks { get; }
     }
 
     public class Disk
     {
-        public Disk(string name, double availableSpace, double totalSpace, string fileSystem, string fileType)
+        public Disk(string name, long availableSpace, long totalSpace, string fileSystem, string fileType)
         {
             this.Name = name;
             this.AvailableSpace = availableSpace;
@@ -35,9 +35,9 @@ namespace Microsoft.Azure.Devices.Edge.Agent.Core.Metrics
 
         public string Name { get; }
 
-        public double AvailableSpace { get; }
+        public long AvailableSpace { get; }
 
-        public double TotalSpace { get; }
+        public long TotalSpace { get; }
 
         public string FileSystem { get; }
